Configure package.json in packageFiles
By explicitly listing `packageFiles` in the `renovate.json` it will help Renovate know not to attempt to parse the template in `src/templates/package.json`. It doesn't really cause you any problem to leave it as-is, because Renovate will skip over any invalid `package.json` files, but it raises an error to me each time and I'd like to get rid of it!